enhancement: more reasonable default enable function
The new default enable function for commit will return true
when `git` is executable and `cwd` is in a git repository.

The new default enable function for github will return true
when `git` and `gh` is executable and remote contains
`github.com`.

When `git` or `gh` is not installed, there will be an warning
when creating the source.

This should fix #15.