fix: completion menu disappears
The disappearance of the completion menu is related with the performance
of CPUs. When you switched between the completion items, if you set
pre-insert, the context will be changed and `blink.cmp` may start a
request to the `lsp`, so it may cause the completion menu to disappear
when your `lsp` is slow. I've tested that when closing the `lsp`,
the completion menu will not disappear.

`blink-cmp-git` has tried to optimize the performance by
removing some unnecessary completions, this should close #24.